requirements.txt
# Core dependencies for Parse Exam Bank project
requests>=2.25.1
beautifulsoup4>=4.9.3
lxml>=4.6.3
selenium>=4.15.0
webdriver-manager>=4.0.0

# Data processing and analysis
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.4.0
seaborn>=0.11.0

# PDF processing
PyPDF2>=3.0.0
pdfplumber>=0.7.0
pymupdf>=1.20.0
Pillow>=10.0.0
pytesseract>=0.3.10
pdf2image>=1.16.0

# Machine Learning and NLP
scikit-learn>=1.0.0
nltk>=3.6.0
spacy>=3.4.0
transformers>=4.20.0
torch>=1.12.0
tokenizers>=0.12.0

# Web development (for future phases)
fastapi>=0.78.0
uvicorn>=0.18.0
pydantic>=1.9.0
jinja2>=3.1.0
streamlit>=1.28.0  # For web UI

# Database
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
redis>=4.3.0

# Development tools
jupyter>=1.0.0
ipython>=8.0.0
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0

# Additional utilities
python-dotenv>=0.19.0
tqdm>=4.64.0
click>=8.0.0
rich>=12.0.0

# Optional: OpenAI API for intelligent extraction
openai>=1.0.0
